seperate all person  methods into one Class
